Crate bevy_a11y

Accessibility for Bevy

Re-exports

pub use accesskit;

Structs

AccessibilityNode

Component to wrap a accesskit::Node, representing this entity to the platform’s accessibility API.

AccessibilityPlugin

Plugin managing non-GUI aspects of integrating with accessibility APIs.

AccessibilityRequested

Resource that tracks whether an assistive technology has requested accessibility information.

ActionRequest

Wrapper struct for accesskit::ActionRequest. Required to allow it to be used as an Event.

Focus

Resource representing which entity has keyboard focus, if any.

ManageAccessibilityUpdates

Resource whose value determines whether the accessibility tree is updated via the ECS.

Enums

AccessibilitySystem

Set enum for the systems relating to accessibility